refactor: delete column from settings
